+++ openwrt-library-arduino/src/protocol/crc.cpp
@@ -9,7 +9,7 @@ uint32_t crc32_ieee(const uint8_t* data, size_t len) {
   uint32_t crc = rpc::RPC_CRC_INITIAL;
   for (size_t i = 0; i < len; i++) {
     crc ^= data[i];
-    for (int j = 0; j < 8; j++) {
+    for (uint8_t j = 0; j < 8; j++) {
       if (crc & 1) {
         crc = (crc >> 1) ^ rpc::RPC_CRC_POLYNOMIAL;
       } else {
